+++ bin/proving-service/src/proxy/worker.rs
@@ -3,6 +3,11 @@ use std::{
     time::{Duration, Instant},
 };
 
+use miden_proving_service::{
+    api::ProverType,
+    error::ProvingServiceError,
+    generated::status::{StatusRequest, status_api_client::StatusApiClient},
+};
 use pingora::lb::Backend;
 use semver::{Version, VersionReq};
 use serde::Serialize;
@@ -10,11 +15,6 @@ use tonic::transport::Channel;
 use tracing::{error, info};
 
 use super::metrics::WORKER_UNHEALTHY;
-use crate::{
-    commands::worker::ProverType,
-    error::ProvingServiceError,
-    generated::status::{StatusRequest, status_api_client::StatusApiClient},
-};
 
 /// The maximum exponent for the backoff.
 ///